+++ tool/common/headers.dart
@@ -1,7 +1,31 @@
-const structFileHeader = '''
+const copyrightHeader = '''
 // Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+''';
+
+const functionsFileHeader = '''
+$copyrightHeader
+
+// Maps FFI prototypes onto the corresponding Win32 API function calls
+
+// THIS FILE IS GENERATED AUTOMATICALLY AND SHOULD NOT BE EDITED DIRECTLY.
+
+// ignore_for_file: unused_import
+
+import 'dart:ffi';
+
+import 'package:ffi/ffi.dart';
+
+import 'callbacks.dart';
+import 'combase.dart';
+import 'guid.dart';
+import 'structs.dart';
+import 'structs.g.dart';
+''';
+
+const structFileHeader = '''
+$copyrightHeader
 
 // Dart representations of common structs used in the Win32 API.
