+++ tool/metadata/generate_win32_tests.dart
@@ -14,8 +14,8 @@ import '../manual_gen/struct_sizes.dart';
 import '../manual_gen/win32api.dart';
 import '../projection/function.dart';
 import '../projection/type.dart';
-import 'generate_win32.dart';
-import 'win32_functions.dart';
+import '../projection/utils.dart';
+import 'generate_win32_functions.dart';
 
 int generateTests(Win32API win32) {
   var testsGenerated = 0;
